GUDjango2/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
dj-static==0.0.6
Django==3.0.6
django-bootstrap4==1.1.1
django-stdimage==5.1.1
gunicorn==20.0.4
mysql==0.0.2
mysqlclient==1.4.6
Pillow==7.1.2
progressbar2==3.51.3
PyMySQL==0.9.3
python-utils==2.4.0
pytz==2020.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
